almost done

+++ main.py
@@ -1,6 +1,7 @@
+import asyncio
 import logging
 import os
-import time
+import shutil
 from collections import defaultdict
 
 import discord
@@ -9,11 +10,9 @@ from discord import FFmpegPCMAudio
 from discord.ext import commands
 from discord.ui import Select, View
 
-import yandex_music_api
 from settings import discord_settings
-from yandex_music_api import get_tracks_info, download_track
-
-import asyncio
+from yandex_music_api import get_tracks_info, get_track_info, get_albums_info, download_track, get_chart_tracks_info, \
+    client
 
 logging.basicConfig(
     filename='events.log',
@@ -23,12 +22,32 @@ logging.basicConfig(
 bot = commands.Bot(command_prefix=discord_settings['command_prefix'],
                    intents=discord.Intents.all())
 
-queue = defaultdict(list)
+
+class QueueList(list):
+    def append(self, __object, ctx) -> None:
+        super().append(__object)
+        on_queue_append(ctx)
+
+
+queue = defaultdict(QueueList)
 to_chose = defaultdict(list)
+
+
+def on_queue_append(ctx):
+    if len(queue[ctx.channel.id]) <= 2:
+        track = queue[ctx.channel.id][-1]
+        download_track(track['id'], path=str(ctx.channel.id) + '/')
 
 
 def mls_to_sm(x):
     return f'{x // 60 // 1000}:{str(x // 1000 - x // 60 // 1000 * 60).rjust(2, "0")}'
+
+
+async def is_user_in_voicechat(ctx):
+    if not ctx.message.author.voice:
+        await ctx.send(ctx.message.author.mention + '\n**Я не могу включить Вам песенку, Вы же не в чате**')
+        return
+    return True if ctx.message.author.voice else False
 
 
 async def play_queue(ctx: discord.ext.commands.context.Context):
@@ -37,9 +56,6 @@ async def play_queue(ctx: discord.ext.commands.context.Context):
     voice = ctx.voice_client
 
     while queue[channel_id]:
-        if len(queue[channel_id]) >= 3:
-            yandex_music_api.download_track(queue[channel_id][2]['id'], path=str(channel_id) + '/')
-
         source = FFmpegPCMAudio(f'{channel_id}/{queue[channel_id][0]["id"]}.mp3', executable=discord_settings[
             'ffmpeg_path'])
 
@@ -48,6 +64,9 @@ async def play_queue(ctx: discord.ext.commands.context.Context):
         await asyncio.sleep(queue[channel_id][0]["duration"] / 1000 + 0.1)
         print('Stopped playing...')
 
+        if not queue[channel_id]:
+            return
+
         if len(queue[channel_id]) >= 2:
             if queue[channel_id][0]["id"] != queue[channel_id][1]["id"]:
                 os.remove(f'{channel_id}/{queue[channel_id][0]["id"]}.mp3')
@@ -55,29 +74,36 @@ async def play_queue(ctx: discord.ext.commands.context.Context):
         else:
             os.remove(f'{channel_id}/{queue[channel_id][0]["id"]}.mp3')
             print('File deleted...')
+
+        if len(queue[channel_id]) >= 3:
+            download_track(queue[channel_id][2]['id'], path=str(channel_id) + '/')
+
         del queue[channel_id][0]
 
 
 @bot.command()
-async def playlist(ctx):  # Информация об очереди
-    embed = discord.Embed(color=0xff9900, title='Очередь из треков')
+async def playlist(ctx):
     channel_id = ctx.channel.id
 
-    embed = discord.Embed(color=0xff9900, title='Очередь из треков')
+    embed = discord.Embed(color=0xff9900, title='Ваш плейлист:')
 
     if len(queue[channel_id]) > 1:
         to_print = ''
         now_playing = ''
-        for n, track in enumerate(queue[channel_id]):
+        for n, track in enumerate(queue[channel_id][:11]):
             if n == 0:
                 now_playing = f"**{track['title']} - {track['artists']}** `Длительность: {mls_to_sm(track['duration'])}`"
             else:
-                to_print += f"**{n}**: **{track['title']} - {track['artists']}** `Длительность: {mls_to_sm(track['duration'])}`\n"
+                to_print += f"**{n}**: **{track['title']} - {track['artists']}** (`Длительность: {mls_to_sm(track['duration'])}`)\n"
         embed.add_field(name='Сейчас играет', value=f"{now_playing}", inline=False)
         embed.add_field(name='В очереди', value=f"{to_print}", inline=False)
 
+        if len(queue[channel_id]) > 10:
+            embed.add_field(name='\n', value='', inline=False)
+            embed.add_field(name='**Всего треков в очереди:**', value=str(len(queue[channel_id])), inline=False)
+
     elif len(queue[channel_id]) == 1:
-        now_playing = f"**{queue[channel_id][0]['title']} - {queue[channel_id][0]['artists']}** `Длительность: {mls_to_sm(queue[channel_id][0]['duration'])}`"
+        now_playing = f"**{queue[channel_id][0]['title']} - {queue[channel_id][0]['artists']}** (`Длительность: {mls_to_sm(queue[channel_id][0]['duration'])}`)"
         embed.add_field(name='Сейчас играет', value=f"{now_playing}", inline=False)
     else:
         embed = discord.Embed(color=0xff9900, title='Я ничего не играю',
@@ -118,36 +144,44 @@ async def skip(ctx: discord.ext.commands.context.Context):
     if not ctx.voice_client:
         embed = discord.Embed(color=0xff9900, title='**В данный момент очередь пуста!**')
         embed.set_footer(**discord_settings['embed_footer'])
-        await ctx.send(embed=embed)
-        print(213)
+        await ctx.send(embed=embed, reference=ctx.message)
         return
 
     ctx.voice_client.stop()
+    await asyncio.sleep(0.1)
     if queue[channel_id]:
+        if len(queue[channel_id]) > 1:
 
-        embed = discord.Embed(color=0xff9900, title='**Вы успешно пропустили трек**')
-        embed.add_field(value=f'**{queue[channel_id][0]["title"]}** - {queue[channel_id][0]["artists"]}',
-                        name=f"**Сейчас играет:**",
-                        inline=False)
-        embed.set_image(url=queue[channel_id][0]['image_url'])
-        embed.set_footer(**discord_settings['embed_footer'])
+            if len(queue[channel_id]) >= 2:
+                if queue[channel_id][0]["id"] != queue[channel_id][1]["id"]:
+                    os.remove(f'{channel_id}/{queue[channel_id][0]["id"]}.mp3')
+                    print('File deleted...')
+            else:
+                os.remove(f'{channel_id}/{queue[channel_id][0]["id"]}.mp3')
+                print('File deleted...')
 
-        await ctx.send(ctx.author.mention, embed=embed)
+            embed = discord.Embed(color=0xff9900, title='**Вы успешно пропустили трек**')
+            embed.add_field(value=f'**{queue[channel_id][1]["title"]}** - {queue[channel_id][1]["artists"]}',
+                            name=f"**Сейчас играет:**",
+                            inline=False)
+            embed.set_image(url=queue[channel_id][1]['image_url'])
+            embed.set_footer(**discord_settings['embed_footer'])
+
+            await ctx.send(ctx.author.mention, embed=embed, reference=ctx.message)
+
+        elif len(queue[channel_id]) == 1:
+            embed = discord.Embed(color=0xff9900, title='**Вы успешно пропустили трек**')
+            embed.add_field(name=f'В данный момент очередь пуста', value='', inline=False)
+
+            await ctx.send(ctx.author.mention, embed=embed, reference=ctx.message)
 
         del queue[channel_id][0]
-
         await play_queue(ctx)
-
     else:
         embed = discord.Embed(color=0xff9900, title='**Вы успешно пропустили трек!**',
                               description='В данный момент очередь пуста')
         embed.set_footer(**discord_settings['embed_footer'])
-        await ctx.send(ctx.author.mention, embed=embed)
-
-
-@bot.command()
-async def s(ctx):
-    await skip(ctx)
+        await ctx.send(ctx.author.mention, embed=embed, reference=ctx.message)
 
 
 @bot.command()
@@ -159,12 +193,12 @@ async def resume(ctx):
 
 @bot.command()
 async def play(ctx, *, name_of_song):
-    channel_id = ctx.channel.id
+    await join(ctx)
 
-    if not ctx.message.author.voice:
-        await ctx.send(ctx.message.author.mention +
-                       '\n**Я не могу включить Вам песенку, Вы же не в чате**')
+    if not await is_user_in_voicechat(ctx):
         return
+
+    channel_id = ctx.channel.id
 
     tracks_info = get_tracks_info(name_of_song)
     message_1 = await ctx.send('***Ищу вашу песенку в Яндекс Музыке...***')
@@ -197,6 +231,10 @@ async def play(ctx, *, name_of_song):
     view.add_item(selectmenu)
 
     async def my_callback(interaction: discord.Interaction):
+        if interaction.user.id != ctx.author.id:
+            await ctx.send(f'{interaction.user.mention}, я не тебя спрашивал!')
+            return
+
         await message.delete()
         await message_1.delete()
         await c(ctx, int(selectmenu.values[0][0]))
@@ -207,18 +245,109 @@ async def play(ctx, *, name_of_song):
 
 
 @bot.command()
-async def p(ctx, *, name_of_song):
-    await play(ctx, name_of_song=name_of_song)
+async def playalbum(ctx, *, name_of_album):
+    await join(ctx)
+
+    if not await is_user_in_voicechat(ctx):
+        return
+
+    channel_id = ctx.channel.id
+
+    albums_info = get_albums_info(name_of_album)
+    message_1 = await ctx.send('***Ищу ваш альбом в Яндекс Музыке...***')
+
+    if not albums_info:
+        await ctx.send(f'{ctx.message.author.mention} К сожалению, я не ничего не нашёл')
+        return
+
+    to_say = ''
+    embed = discord.Embed(color=0xff9900, title='**Я нашел немного альбомов!**')
+    for i, album in enumerate(albums_info, 1):
+        artists = album['artists']
+        to_say += f'**{i}: {album["title"]}** - {artists}  (Треков: {album["track_count"]})\n'
+
+    embed.add_field(name="", value=f"{to_say}", inline=False)
+    embed.set_footer(**discord_settings['embed_footer'])
+
+    selectmenu = Select(options=[
+        discord.SelectOption(label=f'{i}: {album["title"][:50]} - {album["artists"][:40]}') for i, album in
+        enumerate(albums_info, 1)
+    ])
+
+    view = View()
+    view.add_item(selectmenu)
+
+    async def my_callback(interaction: discord.Interaction):
+        if interaction.user.id != ctx.author.id:
+            await ctx.send(f'{ctx.message.author.mention}, я не тебя спрашивал!')
+            return
+
+        await message.delete()
+
+        embed = discord.Embed(color=0xff9900)
+        embed.set_image(url='https://media.tenor.com/YayvrhznSVwAAAAC/yandex-yandexru.gif')
+        await message_1.edit(embed=embed)
+
+        chosed_album = albums_info[int(selectmenu.values[0][0]) - 1]
+
+        tracks = client.albums_with_tracks(chosed_album['id']).volumes[0]
+        for track in tracks:
+            queue[channel_id].append(get_track_info(track), ctx=ctx)
+
+        embed = discord.Embed(color=0xff9900,
+                              title=f'**Вы успешно добавили альбом "{chosed_album["title"]}" в очередь!**',
+                              description=f'**{len(tracks)} треков уже ждут вас**')
+
+        embed.set_image(url=chosed_album['image_url'])
+        embed.set_footer(**discord_settings['embed_footer'])
+        await message_1.delete()
+        await ctx.send(ctx.author.mention, embed=embed)
+        try:
+            await play_queue(ctx)
+        except Exception:
+            pass
+
+    selectmenu.callback = my_callback
+
+    message = await ctx.send(ctx.message.author.mention, embed=embed, view=view)
+
+
+@bot.command()
+async def playchart(ctx):
+    await join(ctx)
+
+    if not await is_user_in_voicechat(ctx):
+        return
+
+    embed = discord.Embed(color=0xff9900)
+    embed.set_image(url='https://media.tenor.com/YayvrhznSVwAAAAC/yandex-yandexru.gif')
+    message = await ctx.send('**Ищем чарты в Яндекс Музыке для вас...**', embed=embed)
+
+    await clear(ctx)
+
+    for track in get_chart_tracks_info():
+        queue[ctx.channel.id].append(track, ctx=ctx)
+
+    embed = discord.Embed(color=0xff9900, title='**Теперь я в режиме чартов!**',
+                          description=f'**Кстати, сейчас в очереди {len(queue[ctx.channel.id])} треков.**')
+
+    track = queue[ctx.channel.id][0]
+    now_playing = f"**{track['title']} - {track['artists']}** (`Длительность: {mls_to_sm(track['duration'])}`)"
+    embed.add_field(name='Первый из них', value=now_playing, inline=False)
+    embed.set_image(url=track['image_url'])
+    embed.set_footer(**discord_settings['embed_footer'])
+    await message.delete()
+    await ctx.send(embed=embed)
+
+    await play_queue(ctx)
 
 
 @bot.command()
 async def playbest(ctx, *, name_of_song):
-    channel_id = ctx.channel.id
-
-    if not ctx.message.author.voice:
-        await ctx.send(ctx.message.author.mention +
-                       '\n**Я не могу включить Вам песенку, Вы же не в чате**')
+    if not await is_user_in_voicechat(ctx):
         return
+
+    channel_id = ctx.channel.id
 
     tracks_info = get_tracks_info(name_of_song, count=1)
     if not tracks_info:
@@ -230,26 +359,20 @@ async def playbest(ctx, *, name_of_song):
 
 
 @bot.command()
-async def pb(ctx, *, name_of_song):
-    await playbest(ctx, name_of_song=name_of_song)
-
-
-@bot.command()
 async def join(ctx):
+    if not await is_user_in_voicechat(ctx):
+        return
+
     if not ctx.voice_client:
         await ctx.message.author.voice.channel.connect()
-        time.sleep(1)
 
 
 async def c(ctx: discord.ext.commands.context.Context, num: int):
-    print('c func started')
-    await join(ctx)
-
     channel_id = ctx.channel.id
 
     num = num - 1
 
-    queue[channel_id].append(to_chose[channel_id][num])
+    queue[channel_id].append(to_chose[channel_id][num], ctx=ctx)
 
     embed = discord.Embed(color=0xff9900,
                           title=f'**{to_chose[channel_id][num]["title"]}** - {to_chose[channel_id][num]["artists"]}')
@@ -260,12 +383,6 @@ async def c(ctx: discord.ext.commands.context.Context, num: int):
         f'{ctx.message.author.mention} добавил(а) `{to_chose[channel_id][num]["title"]} - {to_chose[channel_id][num]["artists"]}` в очередь',
         embed=embed)
 
-    if len(queue[channel_id]) <= 2:
-        if not os.path.exists(str(channel_id)):
-            os.mkdir(str(channel_id))
-
-        download_track(queue[channel_id][len(queue[channel_id]) - 1]['id'], path=str(channel_id) + '/')
-
     try:
         await play_queue(ctx)
     except Exception:
@@ -273,10 +390,20 @@ async def c(ctx: discord.ext.commands.context.Context, num: int):
 
 
 @bot.command()
-async def leave(ctx):  # Ливнуть из войса
+async def clear(ctx):
+    ctx.voice_client.stop()
+    await asyncio.sleep(0.1)
+    shutil.rmtree(str(ctx.channel.id), ignore_errors=True)
+    queue[ctx.channel.id].clear()
+
+
+@bot.command()
+async def leave(ctx):
     if ctx.voice_client:
-        await ctx.guild.voice_client.disconnect()
+        await clear(ctx)
         await ctx.send(f'{ctx.message.author.mention} Пока-пока!')
+        await ctx.guild.voice_client.disconnect()
+
     else:
         await ctx.send('Я не подключен ни к одному из голосовых каналов')
